% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_viewer.R
\name{ambient_viewer}
\alias{ambient_viewer}
\title{Ambient Viewer app}
\usage{
ambient_viewer()
}
\description{
This function launches the Ambient Viewer app, a Shiny application for visualizing and analyzing sleep data.
}
