% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_epochs}
\alias{clean_epochs}
\title{Clean epoch data}
\usage{
clean_epochs(epochs)
}
\arguments{
\item{epochs}{A dataframe containing the epoch data}
}
\value{
The cleaned-up epochs dataframe
}
\description{
Clean epoch data
}
\examples{
epochs <- clean_epochs(AmbientViewer::example_epochs)
}
