% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{composite_phase_deviation}
\alias{composite_phase_deviation}
\title{Calculate Composite Phase Deviation (CPD)}
\usage{
composite_phase_deviation(sessions, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions data frame}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{time_at_midsleep}
\item \code{is_workday}
\item \code{night}
}}
}
\value{
The Composite Phase Deviation (CPD) value
}
\description{
This function calculates the Composite Phase Deviation (CPD) metric, used to measure
the regularity of the sleep patterns.
}
\examples{
composite_phase_deviation(example_sessions)
}
\seealso{
Other sleep metrics: 
\code{\link{chronotype}()},
\code{\link{interdaily_stability}()},
\code{\link{sleep_regularity_index}()},
\code{\link{social_jet_lag}()}
}
\concept{sleep metrics}
