% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stats.R
\name{convert_angle_to_time}
\alias{convert_angle_to_time}
\title{Convert an angle to time}
\usage{
convert_angle_to_time(angle, unit = "second")
}
\arguments{
\item{angle}{A numeric value representing the angle in radians.}

\item{unit}{A string indicating the unit of time. Can be "second", "minute", or "hour".}
}
\value{
A numeric value representing the time in the specified unit.
}
\description{
This function converts an angle in radians to time in the provided unit (can be "second", "minute" or "hour").
}
