% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stats.R
\name{convert_times_to_mean_angle}
\alias{convert_times_to_mean_angle}
\title{Convert a vector of times to a mean angle}
\usage{
convert_times_to_mean_angle(times, unit = "second")
}
\arguments{
\item{times}{A vector of times in seconds.}

\item{unit}{A string indicating the unit of time. Can be "second", "minute", or "hour".}
}
\value{
A numeric value representing the mean angle in radians.
}
\description{
This function converts a vector of times to a mean angle in radians.
It is useful to calculate average times spanning midnight
}
