% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_epochs}
\alias{example_epochs}
\title{Example Epoch data}
\format{
\subsection{\code{example_epochs}}{

A data frame with 18,755 rows and 15 columns.
Each row represents a time-point (or epoch) in a session. Epochs are 30 seconds long.
The columns are as follows:
\itemize{
\item timestamp: The time at which the epoch was recorded in UTC.
\item subject_id: The ID of the subject.
\item signal_quality_mean: The mean signal quality of the epoch.
\item movement_fast_mean: The mean movement detected during the epoch.
\item movement_fast_nonzero_pct
\item distance_mean: the distance of the subject from the device in meters.
\item motion_data_count: The number of data points in the epoch (30).
\item light_ambient_mean: The ambient light level during the epoch.
\item sound_amplitude_mean: The sound amplitude during the epoch.
\item temperature_ambient_mean: The ambient temperature during the epoch.
\item humidity_mean: The ambient humidity during the epoch.
\item pressure_mean: The ambient pressure during the epoch.
\item indoor_air_quality_mean: The indoor air quality during the epoch.
\item epoch_duration: The precise duration of the epoch (seconds).
\item sleep_stage: The sleep stage as established with the VT algorithm. They are encoded as numbers 0-5
}
}
}
\source{
data-raw/example_epochs.csv
}
\usage{
example_epochs
}
\description{
A data frame containing epoch data recorded by a Somnofy device.
}
\keyword{datasets}
