% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sessions}
\alias{example_sessions}
\title{Example Sessions data}
\format{
\subsection{\code{example_sessions}}{

A data frame with 124 rows and 60 columns.
Each row represents a session. Columns contain metadata about the session, including:
\itemize{
\item start_time: The start time of the session in UTC.
\item end_time: The end time of the session in UTC.
\item subject_id: The ID of the subject.
\item device_serial_number: The serial number of the device used.
\item time_at_sleep_onset: The time at which the subject fell asleep.
\item time_at_wakeup: The time at which the subject woke up.
Columns also include various metrics averaged over the session, such as:
\item mean heart rate
\item mean respiration rate
Finally, some columns contain environmental parameters, such as:
\item Temperature
\item Humidity
\item Light intensity
\item Noise level
\item Atmospheric pressure
}
}
}
\source{
data-raw/example_sessions.csv
}
\usage{
example_sessions
}
\description{
A data frame containing sessions recorded by a Somnofy device.
}
\keyword{datasets}
