% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sessions_v1}
\alias{example_sessions_v1}
\title{Example Sessions data (Somnofy API v1)}
\format{
\subsection{\code{example_sessions_v1}}{

A data frame with 87 rows and 70 columns.
Each row represents a session. Columns contain metadata about the session, including:
\itemize{
\item user_id: The ID of the recorded subject.
\item sex: The sex of the recorded subject.
\item birth_year: The year of birth of the recorded subject.
\item session_start: The start time of the session in UTC.
\item session_end: The end time of the session in UTC.
\item time_at_sleep_onset: The time at which the subject fell asleep.
\item time_at_wakeup: The time at which the subject woke up.
Columns also include various metrics averaged over the session, such as:
\item mean heart rate
\item mean respiration rate
Finally, some columns contain environmental parameters, such as:
\item Temperature
\item Humidity
\item Light intensity
\item Noise level
\item Atmospheric pressure
}
}
}
\source{
data-raw/example_sessions_v1.csv
}
\usage{
example_sessions_v1
}
\description{
A data frame containing sessions recorded by a Somnofy device.
}
\keyword{datasets}
