% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_by_age_range}
\alias{filter_by_age_range}
\title{Filter sessions by age range}
\usage{
filter_by_age_range(sessions, min_age, max_age, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{min_age}{The minimum age of the subjects (inclusive)}

\item{max_age}{The maximum age of the subjects (inclusive)}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{birth_year}
}}
}
\value{
The sessions dataframe with only the sessions that belong to subjects within the specified age range
}
\description{
Filter sessions by age range
}
\examples{
filtered_sessions <- filter_by_age_range(example_sessions, min_age = 11, max_age = 18)
}
\seealso{
Other filtering: 
\code{\link{filter_by_night_range}()},
\code{\link{filter_by_sex}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{remove_sessions_no_sleep}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_min_time_in_bed}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
