% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_duplicate_sessions}
\alias{get_duplicate_sessions}
\title{Get duplicate sessions}
\usage{
get_duplicate_sessions(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe.}
}
\value{
A dataframe containing only the duplicate sessions.
}
\description{
This function identifies sessions that occurred during the same "night" (from 12pm to 12pm).
}
\examples{
duplicate_sessions <- get_duplicate_sessions(sessions)
}
