% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_epochs_summary}
\alias{get_epochs_summary}
\title{Summarise epoch information}
\usage{
get_epochs_summary(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A single-row dataframe summarising epoch information
}
\description{
This function displays the number of sessions in the epoch data, as well as the start and end dates of the epoch data
}
\examples{
get_epochs_summary(example_epochs)
}
