% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_sessions_summary}
\alias{get_sessions_summary}
\title{Make a summary of session information}
\usage{
get_sessions_summary(sessions, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions dataframe.}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{time_in_bed}
\item \code{sleep_period}
}}
}
\value{
A single-row dataframe summarizing session information.
}
\description{
This function summarises session information, including the number of sessions, mean session length,
mean time at sleep onset and wakeup, subject and device ID.
}
\examples{
get_sessions_summary(example_sessions)
}
\seealso{
\code{\link[=get_epochs_summary]{get_epochs_summary()}} to summarise epoch information.

Other data tables: 
\code{\link{get_epochs_summary}()},
\code{\link{get_non_complying_sessions}()},
\code{\link{get_removed_sessions}()}
}
\concept{data tables}
