% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stats.R
\name{get_time_per_day}
\alias{get_time_per_day}
\title{Get the number of seconds, minutes or hours in a day}
\usage{
get_time_per_day(unit = "second")
}
\arguments{
\item{unit}{A string indicating the unit of time. Can be "second", "minute", or "hour".}
}
\value{
The number of seconds, minutes or hours in a day.
}
\description{
Get the number of seconds, minutes or hours in a day
}
