% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{group_epochs_by_night}
\alias{group_epochs_by_night}
\title{Create a grouping by night for epoch data}
\usage{
group_epochs_by_night(epochs, col_names = NULL)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{timestamp}
}}
}
\value{
The epochs dataframe with the \code{night} column added
}
\description{
Create a grouping by night for epoch data
}
\details{
The function creates a new column \code{night} that groups the epochs by night.
Timepoints before 12 PM are considered part of the previous night.
}
\examples{
epochs <- group_epochs_by_night(example_epochs)
}
\seealso{
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}} to group session data by night.

Other time processing: 
\code{\link{convert_angle_to_time}()},
\code{\link{convert_times_to_mean_angle}()},
\code{\link{group_sessions_by_night}()},
\code{\link{max_time}()},
\code{\link{mean_time}()},
\code{\link{min_time}()},
\code{\link{shift_times_by_12h}()}
}
\concept{time processing}
