% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{interdaily_stability}
\alias{interdaily_stability}
\title{Calculate Interdaily Stability (IS)}
\usage{
interdaily_stability(epochs, col_names = NULL)
}
\arguments{
\item{epochs}{The epochs data frame}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{is_asleep}
}}
}
\value{
The Interdaily Stability (IS) value
}
\description{
This function calculates the Interdaily Stability (IS) metric from a binary awake/asleep variable
}
\examples{
interdaily_stability(example_epochs)
}
\seealso{
Other sleep metrics: 
\code{\link{chronotype}()},
\code{\link{composite_phase_deviation}()},
\code{\link{sleep_regularity_index}()},
\code{\link{social_jet_lag}()}
}
\concept{sleep metrics}
