% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{is_iso8601_datetime}
\alias{is_iso8601_datetime}
\title{Check if a column contains datetime data in ISO 8601 format}
\usage{
is_iso8601_datetime(column)
}
\arguments{
\item{column}{A vector of character strings to check}
}
\value{
TRUE if all non-NA values are in ISO 8601 format, FALSE otherwise
}
\description{
Check if a column contains datetime data in ISO 8601 format
}
