% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load session and epoch data}
\usage{
load_data(folder, basename)
}
\arguments{
\item{folder}{The folder where the data is stored}

\item{basename}{The time range of the files to load}
}
\value{
A list containing the sessions and epochs dataframes
}
\description{
Load session and epoch data
}
