% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_sessions_batch}
\alias{load_sessions_batch}
\title{Load session data in batch mode}
\usage{
load_sessions_batch(folder_path, pattern = "")
}
\arguments{
\item{folder_path}{The path to the folder containing session files}

\item{pattern}{An optional pattern to filter files in the folder}
}
\value{
A dataframe containing the combined session data from all files in the folder
}
\description{
Load session data in batch mode
}
\seealso{
Other data loading: 
\code{\link{load_epochs}()},
\code{\link{load_sessions}()},
\code{\link{read_edf_epochs}()},
\code{\link{read_edf_sessions}()}
}
\concept{data loading}
