% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stats.R
\name{mean_time}
\alias{mean_time}
\title{Calculate the mean time from a vector of time strings}
\usage{
mean_time(time_vector)
}
\arguments{
\item{time_vector}{A vector of time strings in the format "YYYY-MM-DD HH:MM:SS".}
}
\value{
A string representing the mean time in the format "HH:MM".
}
\description{
This function calculates the mean time from a vector of time strings in the format "YYYY-MM-DD HH:MM:SS".
}
