% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{min_time}
\alias{min_time}
\title{Calculate the minimum time from 12pm to 12pm}
\usage{
min_time(time_vector)
}
\arguments{
\item{time_vector}{A vector of time strings in the format "YYYY-MM-DD HH:MM:SS".}
}
\value{
A string representing the minimum time in the format "HH:MM".
}
\description{
This function calculates the minimum time from a vector of time strings in the format "YYYY-MM-DD HH:MM:SS".
It considers a time window from 12pm to 12pm the next day, so 11:00 is considered later than 13:00.
}
\seealso{
Other time processing: 
\code{\link{convert_angle_to_time}()},
\code{\link{convert_times_to_mean_angle}()},
\code{\link{group_epochs_by_night}()},
\code{\link{group_sessions_by_night}()},
\code{\link{max_time}()},
\code{\link{mean_time}()},
\code{\link{shift_times_by_12h}()}
}
\concept{time processing}
