% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtimes_waketimes.R
\name{plot_bedtimes_waketimes}
\alias{plot_bedtimes_waketimes}
\title{Plot bedtimes and waketimes}
\usage{
plot_bedtimes_waketimes(sessions, groupby = "night")
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{groupby}{The grouping variable for the plot. Can be "night", "workday", or "weekday".}
}
\value{
A ggplot graph showing the bedtimes and waketimes
}
\description{
Plot bedtimes and waketimes
}
