% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_hypnogram}
\alias{plot_hypnogram}
\title{Plot Hypnogram}
\usage{
plot_hypnogram(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A ggplot object showing the hypnogram as bars
}
\description{
Plot Hypnogram
}
\seealso{
\code{\link[=plot_sleep_stages]{plot_sleep_stages()}} to show the proportion of each sleep stage per day

Other plot epochs: 
\code{\link{plot_sleep_spiral}()},
\code{\link{plot_sleep_stages}()},
\code{\link{plot_timeseries}()}
}
\concept{plot epochs}
