% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_clock.R
\name{plot_sleep_clock}
\alias{plot_sleep_clock}
\title{Plot Sleep Clock}
\usage{
plot_sleep_clock(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe}
}
\value{
A ggplot object showing the sleep clock
}
\description{
Plot Sleep Clock
}
