% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_spiral.R
\name{plot_sleep_spiral}
\alias{plot_sleep_spiral}
\title{Plot Sleep Spiral}
\usage{
plot_sleep_spiral(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A ggplot object showing the sleep spiral
}
\description{
Plot Sleep Spiral
}
\seealso{
Other plot epochs: 
\code{\link{plot_hypnogram}()},
\code{\link{plot_sleep_stages}()},
\code{\link{plot_timeseries}()}
}
\concept{plot epochs}
