% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series data for a given variable and time window}
\usage{
plot_timeseries(epochs, variable, start_date, end_date)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{variable}{The variable to plot (e.g., "temperature_ambient_mean")}

\item{start_date}{The start date for the time window (YYYY-MM-DD)}

\item{end_date}{The end date for the time window (YYYY-MM-DD)}
}
\value{
A ggplot object
}
\description{
Plot time series data for a given variable and time window
}
\examples{
plot <- plot_timeseries(epochs, "temperature_ambient_mean", "2025-03-04", "2025-03-11")
}
