% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_timeseries_sessions}
\alias{plot_timeseries_sessions}
\title{Plot session time series data for a given variable}
\usage{
plot_timeseries_sessions(sessions, variable, exclude_zero = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{variable}{The variable to plot (e.g., "time_at_sleep_onset")}

\item{exclude_zero}{Logical, whether to exclude zero values from the plot (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Plot session time series data for a given variable
}
\examples{
plot <- plot_timeseries(sessions, "temperature_ambient_mean", exclude_zero = TRUE)
}
