% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_reading.R
\name{read_edf_epochs}
\alias{read_edf_epochs}
\title{Read EDF Epochs}
\usage{
read_edf_epochs(file)
}
\arguments{
\item{file}{The path to the EDF file}

\item{timestamp}{The name of the timestamp column in the EDF annotations signal. Default is "timestamp".}

\item{annotation}{The name of the annotation column in the EDF annotations signal. Default is "annotation".}
}
\value{
A dataframe containing the epoch data extracted from the EDF file signals
}
\description{
Read EDF Epochs
}
\details{
The function reads the signals of the EDF file to extract epoch information. It must contain
a column for timestamps and a column for sleep stage annotations.
}
\seealso{
Other data loading: 
\code{\link{load_batch}()},
\code{\link{load_epochs}()},
\code{\link{load_sessions}()},
\code{\link{read_edf_sessions}()}
}
\concept{data loading}
