% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_reading.R
\name{read_edf_sessions}
\alias{read_edf_sessions}
\title{Read EDF Sessions}
\usage{
read_edf_sessions(file)
}
\arguments{
\item{file}{The path to the EDF file}
}
\value{
A dataframe containing the session data extracted from the EDF file header
}
\description{
Read EDF Sessions
}
\details{
The function reads the header of the EDF file to extract session information such as
start time, duration, and calculates end time and midsleep time.
}
\seealso{
Other data loading: 
\code{\link{load_epochs}()},
\code{\link{load_sessions}()}
}
\concept{data loading}
