% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{remove_sessions_no_sleep}
\alias{remove_sessions_no_sleep}
\title{Remove sessions with no sleep}
\usage{
remove_sessions_no_sleep(sessions, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions have a sleep period greater than 0}
}
\value{
The sessions dataframe with only the sessions that have a sleep period greater than 0, or a logical vector if \code{return_mask} is TRUE
}
\description{
Remove sessions with no sleep
}
\details{
This function uses columns:
\itemize{
\item \code{sleep_period}
}
}
\examples{
filtered_sessions <- remove_sessions_no_sleep(example_sessions)
}
\seealso{
Other filtering: 
\code{\link{filter_by_age_range}()},
\code{\link{filter_by_night_range}()},
\code{\link{filter_by_sex}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_min_time_in_bed}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
