% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{sd_time}
\alias{sd_time}
\title{Calculate the circular standard deviation of a vector of times}
\usage{
sd_time(time_vector, unit = "hour")
}
\arguments{
\item{time_vector}{A vector of time strings in format "YYYY-MM-DD HH:MM:SS", "HH:MM:SS" or "HH:MM".}

\item{unit}{The unit of time for the result. Can be "second", "minute", or "hour". Default is "hour".}
}
\value{
A numeric value representing the standard deviation in the specified unit.
}
\description{
This function calculates the standard deviation of a vector of time strings,
accounting for the circular nature of time (e.g., 23:59 is close to 00:00).
}
\examples{
sd_time(c("23:59", "00:01"))
}
\seealso{
Other time processing: 
\code{\link{group_epochs_by_night}()},
\code{\link{group_sessions_by_night}()},
\code{\link{max_time}()},
\code{\link{mean_time}()},
\code{\link{min_time}()},
\code{\link{shift_times_by_12h}()},
\code{\link{time_diff}()}
}
\concept{time processing}
