% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{select_subjects}
\alias{select_subjects}
\title{Select subjects by ID}
\usage{
select_subjects(sessions, subject_ids, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{subject_ids}{The subject IDs to select}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions belong to the specified subjects
\itemize{
\item \code{subject_id}
}}
}
\value{
The sessions dataframe with only the sessions that belong to the specified subjects, or a logical vector if \code{return_mask} is TRUE
}
\description{
Select subjects by ID
}
\details{
This function uses columns:
}
\examples{
filtered_sessions <- select_subjects(example_sessions, c("sub_01JNDH3Z5NP0PSV82NFBGPV31X"))
}
\seealso{
\code{\link[=select_devices]{select_devices()}} to select sessions by device ID.

Other filtering: 
\code{\link{filter_by_age_range}()},
\code{\link{filter_by_night_range}()},
\code{\link{filter_by_sex}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{remove_sessions_no_sleep}()},
\code{\link{select_devices}()},
\code{\link{set_min_time_in_bed}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
