% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{set_data_type}
\alias{set_data_type}
\title{Set the data type for a dataframe}
\usage{
set_data_type(df, data_type)
}
\arguments{
\item{df}{The dataframe to set the data type for}

\item{data_type}{The data type to set. Currently available data types: "somnofy_v1", "somnofy_v2"}
}
\value{
The dataframe with the data type set
}
\description{
Set the data type for a dataframe
}
\details{
The dataframe type is used by Ambient Viewer functions to determine the correct column names.
Note: you do not need to set the data type if you are using the \code{load_sessions} or \code{load_epochs} functions.
}
\examples{
example_sessions <- set_data_type(example_sessions, "somnofy_v2")
}
