% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_min_time_in_bed}
\alias{set_min_time_in_bed}
\title{Set minimum time in bed}
\usage{
set_min_time_in_bed(sessions, min_time_in_bed)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{min_time_in_bed}{The minimum time in bed in hours}
}
\value{
The sessions dataframe with only the sessions that meet the minimum time in bed requirement
}
\description{
Set minimum time in bed
}
