% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_session_sleep_onset_range}
\alias{set_session_sleep_onset_range}
\title{Set sleep onset time range}
\usage{
set_session_sleep_onset_range(sessions, from_time, to_time)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{from_time}{Include sessions where sleep started after this time (in format HH:MM)}

\item{to_time}{Include sessions where sleep started before this time (in format HH:MM)}
}
\value{
The sessions dataframe with only the sessions where sleep started within the specified time range
}
\description{
Set sleep onset time range
}
