% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_distributions.R
\name{sleeptimes_histogram}
\alias{sleeptimes_histogram}
\title{Plot histograms for sleep onset, midsleep, and wakeup times}
\usage{
sleeptimes_histogram(sessions, col_names = NULL, binwidth = 0.25)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{time_at_midsleep}
}}

\item{binwidth}{The width of the bins for the histogram (default 0.25)}
}
\value{
A ggplot object with three overlaid histograms (sleep onset, midsleep, wakeup)
}
\description{
Plot histograms for sleep onset, midsleep, and wakeup times
}
