% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{social_jet_lag}
\alias{social_jet_lag}
\title{Calculate Social Jet Lag}
\usage{
social_jet_lag(sessions, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions data frame}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{time_at_midsleep}
\item \code{is_workday}
}}
}
\value{
The Social Jet Lag (SJL) value in hours
}
\description{
This function calculates the Social Jet Lag (SJL) metric as the difference in mid-sleep times
between workdays and free days.
}
\examples{
social_jet_lag(example_sessions)
}
\seealso{
Other sleep metrics: 
\code{\link{chronotype}()},
\code{\link{composite_phase_deviation}()},
\code{\link{interdaily_stability}()},
\code{\link{sleep_regularity_index}()}
}
\concept{sleep metrics}
