% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_helpers.R
\name{write_log}
\alias{write_log}
\title{write_log}
\usage{
write_log(logger, ..., type = "default")
}
\arguments{
\item{logger}{The logger to write the text to. Can be \code{NULL} or a function}

\item{...}{Messages to write to the logger}

\item{type}{One of \code{default}, \code{info}, \code{error}, \code{warning}, \code{starting}, \code{complete}}
}
\value{
No return value, called for side effects
}
\description{
For internal use. Add text to a logger
}
\keyword{internal}
