% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_epochs}
\alias{clean_epochs}
\title{Clean epoch data}
\usage{
clean_epochs(epochs, col_names = NULL)
}
\arguments{
\item{epochs}{A dataframe containing the epoch data}

\item{col_names}{A named list mapping standard column names to the column names in the epochs dataframe}
}
\value{
A list with two elements:
\itemize{
\item epochs: the cleaned-up epochs dataframe
\item col: the updated column name mapping
}
}
\description{
Clean epoch data
}
\examples{
res <- clean_epochs(AmbientViewer::example_epochs)
epochs <- res$epochs
col_names <- res$col
}
