% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{convert_angle_to_time}
\alias{convert_angle_to_time}
\title{Convert an angle to time}
\usage{
convert_angle_to_time(angle, unit = "second")
}
\arguments{
\item{angle}{A numeric value representing the angle in radians.}

\item{unit}{A string indicating the unit of time. Can be "second", "minute", or "hour".}
}
\value{
A numeric value representing the time in the specified unit.
}
\description{
This function converts an angle in radians to time in the provided unit (can be "second", "minute" or "hour").
}
\examples{
convert_angle_to_time(pi/2, unit = "hour")
}
\seealso{
\code{\link[=convert_times_to_mean_angle]{convert_times_to_mean_angle()}} to calculate the average angle from a vector of time values.

Other time processing: 
\code{\link{convert_times_to_mean_angle}()},
\code{\link{group_epochs_by_night}()},
\code{\link{group_sessions_by_night}()},
\code{\link{max_time}()},
\code{\link{mean_time}()},
\code{\link{min_time}()},
\code{\link{shift_times_by_12h}()},
\code{\link{time_diff}()}
}
\concept{time processing}
