% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{convert_times_to_mean_angle}
\alias{convert_times_to_mean_angle}
\title{Convert a vector of times to a mean angle}
\usage{
convert_times_to_mean_angle(times, unit = "second")
}
\arguments{
\item{times}{A vector of times in seconds.}

\item{unit}{A string indicating the unit of time. Can be "second", "minute", or "hour".}
}
\value{
A numeric value representing the mean angle in radians.
}
\description{
This function converts a vector of times to a mean angle in radians.
It is useful to calculate average times spanning midnight
}
\examples{
convert_times_to_mean_angle(c(23, 10, 0), unit = "hour")
}
\seealso{
\code{\link[=convert_angle_to_time]{convert_angle_to_time()}} to convert the mean angle back to time format.

Other time processing: 
\code{\link{convert_angle_to_time}()},
\code{\link{group_epochs_by_night}()},
\code{\link{group_sessions_by_night}()},
\code{\link{max_time}()},
\code{\link{mean_time}()},
\code{\link{min_time}()},
\code{\link{shift_times_by_12h}()}
}
\concept{time processing}
