% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_epochs_from_sessions}
\alias{filter_epochs_from_sessions}
\title{Filter epochs based on session IDs}
\usage{
filter_epochs_from_sessions(epochs, sessions)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{sessions}{The sessions dataframe}
}
\value{
The epochs dataframe with only the epochs that belong to the specified sessions
}
\description{
Filter epochs based on session IDs
}
