% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{get_non_complying_sessions}
\alias{get_non_complying_sessions}
\title{Get non-complying sessions (i.e. where there is more than one session on the same day)}
\usage{
get_non_complying_sessions(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe}
}
\value{
The sessions dataframe with only the sessions that are non-complying
}
\description{
Get non-complying sessions (i.e. where there is more than one session on the same day)
}
\details{
This function uses columns:
\itemize{
\item \code{night}
}
}
\examples{
duplicate_sessions <- get_non_complying_sessions(example_sessions)
}
\seealso{
Other data tables: 
\code{\link{get_epochs_summary}()},
\code{\link{get_removed_sessions}()},
\code{\link{get_sessions_summary}()}
}
\concept{data tables}
