% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{get_removed_sessions}
\alias{get_removed_sessions}
\title{Get a table of sessions that were removed during filtering}
\usage{
get_removed_sessions(sessions, filtered_sessions, col_names = NULL)
}
\arguments{
\item{sessions}{The original sessions dataframe}

\item{filtered_sessions}{The filtered sessions dataframe}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{id}
\item \code{sleep_period}
}}
}
\value{
The sessions dataframe with only the sessions that were removed during filtering
}
\description{
Get a table of sessions that were removed during filtering
}
\examples{
filtered_sessions <- set_session_start_time_range(example_sessions, "22:00", "06:00")
removed_sessions <- get_removed_sessions(example_sessions, filtered_sessions)
}
\seealso{
Other data tables: 
\code{\link{get_epochs_summary}()},
\code{\link{get_non_complying_sessions}()},
\code{\link{get_sessions_summary}()}
}
\concept{data tables}
