% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_sessions_summary}
\alias{get_sessions_summary}
\title{Make a summary of session information}
\usage{
get_sessions_summary(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe.}
}
\value{
A single-row dataframe summarizing session information.
}
\description{
This function summarizes session information, including the number of sessions, mean session length,
mean time at sleep onset and wakeup, subject and device ID.
}
\examples{
get_sessions_summary(example_sessions)
}
