% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{group_epochs_by_night}
\alias{group_epochs_by_night}
\title{Create a grouping by night for epoch data}
\usage{
group_epochs_by_night(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
The epochs dataframe with the \code{night} column added
}
\description{
Create a grouping by night for epoch data
}
\details{
The function creates a new column \code{night} that groups the epochs by night,
and an \code{adjusted_time} column to facilitate plotting from 12PM to 12PM.
Timepoints before 12 PM are considered part of the previous night.
}
\examples{
epochs <- group_epochs_by_night(epochs)
}
