% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{interdaily_stability}
\alias{interdaily_stability}
\title{Calculate Interdaily Stability (IS)}
\usage{
interdaily_stability(epochs, variable, time_col = "timestamp")
}
\arguments{
\item{epochs}{The epochs data frame}

\item{variable}{The name of the binary sleep variable, where 0 = awake and 1 = asleep}

\item{time_col}{The name of the column containing timestamps (default is "timestamp")}
}
\value{
The Interdaily Stability (IS) value
}
\description{
This function calculates the Interdaily Stability (IS) metric from a binary awake/asleep variable
}
\examples{
interdaily_stability(example_epochs, "is_asleep")
}
\seealso{
Other sleep metrics: 
\code{\link{social_jet_lag}()}
}
\concept{sleep metrics}
