% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_epochs}
\alias{load_epochs}
\title{Load epoch data}
\usage{
load_epochs(epochs_file)
}
\arguments{
\item{epochs_file}{The path to the epochs file}
}
\value{
A dataframe containing the epoch data
}
\description{
Load epoch data
}
\details{
The function loads the epoch data from a CSV file and groups the epochs by night.
}
