% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actigram.R
\name{plot_actigram}
\alias{plot_actigram}
\title{Plot an Actigram}
\usage{
plot_actigram(epochs)
}
\arguments{
\item{epochs}{The epochs data frame}
}
\value{
A ggplot object representing the actigram
}
\description{
Generate an actigram from the Somnofy epoch data.
}
\details{
This function uses columns:
\itemize{
\item \code{sleep_period}
\item \code{signal_quality_mean}
\item \code{sleep_stage}
}
}
