% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtimes_waketimes.R
\name{plot_bedtimes_waketimes}
\alias{plot_bedtimes_waketimes}
\title{Plot bedtimes and waketimes}
\usage{
plot_bedtimes_waketimes(sessions, groupby = "night", color_by = "default")
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{groupby}{The grouping variable for the plot. Can be "night", "workday", or "weekday".}

\item{color_by}{The variable to color the bars by. Can be "default" or any other column name in the sessions dataframe.
Note that if color_by is anything else than "default", groupby will be set to "night".}
}
\value{
A ggplot graph showing the bedtimes and waketimes
}
\description{
Plot bedtimes and waketimes
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{is_workday}
}
}
