% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_bubbles.R
\name{plot_sleep_bubbles}
\alias{plot_sleep_bubbles}
\title{Plot Sleep Bubbles}
\usage{
plot_sleep_bubbles(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe.}
}
\value{
A ggplot object containing the sleep bubbles graph.
}
\description{
This function creates a bubble plot of sleep sessions, where the size and colour of the bubbles represents the sleep duration.
}
