% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_spiral.R
\name{plot_sleep_spiral}
\alias{plot_sleep_spiral}
\title{Plot Sleep Spiral}
\usage{
plot_sleep_spiral(epochs, color_by = "default", col_names = NULL)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{color_by}{The variable to color the spiral by. Can be "default" or any other column name in the epochs dataframe.}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{is_asleep}
}}
}
\value{
A ggplot object showing the sleep spiral
}
\description{
Plot Sleep Spiral
}
\seealso{
Other plot epochs: 
\code{\link{plot_hypnogram}()},
\code{\link{plot_sleep_stages}()},
\code{\link{plot_timeseries}()}
}
\concept{plot epochs}
