% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_stages.R
\name{plot_sleep_stages}
\alias{plot_sleep_stages}
\title{Plot Sleep Stages}
\usage{
plot_sleep_stages(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A ggplot object showing the proportion of sleep stages for each night
}
\description{
Plot Sleep Stages
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{sleep_stage}
}
}
\seealso{
\code{\link[=plot_hypnogram]{plot_hypnogram()}} to show the detailed sleep stages over time

Other plot epochs: 
\code{\link{plot_hypnogram}()},
\code{\link{plot_sleep_spiral}()},
\code{\link{plot_timeseries}()}
}
\concept{plot epochs}
