% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot epoch time series data for a given variable}
\usage{
plot_timeseries(epochs, variable, exclude_zero = FALSE)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{variable}{The variable to plot (e.g., "temperature_ambient_mean")}

\item{exclude_zero}{Logical, whether to exclude zero values from the plot (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Plot epoch time series data for a given variable
}
