% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot epoch time series data for a given variable}
\usage{
plot_timeseries(epochs, variable, color_by = "default", exclude_zero = FALSE)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{variable}{The variable to plot (e.g., "temperature_ambient_mean")}

\item{color_by}{The variable to color the points by. Can be "default" or any other column name in the epochs dataframe.}

\item{exclude_zero}{Logical, whether to exclude zero values from the plot (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Plot epoch time series data for a given variable
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{night}
}
}
\seealso{
\code{\link[=plot_timeseries_sessions]{plot_timeseries_sessions()}} to plot session data.

Other plot epochs: 
\code{\link{plot_hypnogram}()},
\code{\link{plot_sleep_spiral}()},
\code{\link{plot_sleep_stages}()}
}
\concept{plot epochs}
