% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_timeseries_sessions}
\alias{plot_timeseries_sessions}
\title{Plot session time series data for a given variable}
\usage{
plot_timeseries_sessions(
  sessions,
  variable,
  color_by = "default",
  exclude_zero = FALSE
)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{variable}{The variable to plot (e.g., "time_at_sleep_onset")}

\item{color_by}{The variable to color the points by. Can be "default" or any other column name in the sessions dataframe.}

\item{exclude_zero}{Logical, whether to exclude zero values from the plot (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Plot session time series data for a given variable
}
\details{
This function uses columns:
\itemize{
\item \code{night}
}
}
\seealso{
\code{\link[=plot_timeseries]{plot_timeseries()}} to plot epoch data.

Other plot sessions: 
\code{\link{plot_sleep_bubbles}()},
\code{\link{plot_sleep_clock}()}
}
\concept{plot sessions}
