% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{remove_sessions_no_sleep}
\alias{remove_sessions_no_sleep}
\title{Remove sessions with no sleep}
\usage{
remove_sessions_no_sleep(sessions, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{sleep_period}
}}
}
\value{
The sessions dataframe with only the sessions that have a sleep period greater than 0
}
\description{
Remove sessions with no sleep
}
\examples{
filtered_sessions <- remove_sessions_no_sleep(example_sessions)
}
\seealso{
Other filtering: 
\code{\link{filter_by_night_range}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_min_time_in_bed}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
