% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{remove_sessions_no_sleep}
\alias{remove_sessions_no_sleep}
\title{Remove sessions with no sleep}
\usage{
remove_sessions_no_sleep(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe}
}
\value{
The sessions dataframe with only the sessions that have a sleep period greater than 0
}
\description{
Remove sessions with no sleep
}
\examples{
sessions <- remove_sessions_no_sleep(sessions)
}
