% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{select_devices}
\alias{select_devices}
\title{Select devices by ID}
\usage{
select_devices(sessions, device_ids)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{device_ids}{The device IDs to select}
}
\value{
The sessions dataframe with only the sessions recorded by the specified devices
}
\description{
Select devices by ID
}
\seealso{
\code{\link[=select_subjects]{select_subjects()}} to select sessions by subject ID.
}
