% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{select_subjects}
\alias{select_subjects}
\title{Select subjects by ID}
\usage{
select_subjects(sessions, subject_ids)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{subject_ids}{The subject IDs to select}
}
\value{
The sessions dataframe with only the sessions that belong to the specified subjects
}
\description{
Select subjects by ID
}
\seealso{
\code{\link[=select_devices]{select_devices()}} to select sessions by device ID.
}
