% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_min_time_in_bed}
\alias{set_min_time_in_bed}
\title{Set minimum time in bed}
\usage{
set_min_time_in_bed(sessions, min_time_in_bed, col_names = NULL)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{min_time_in_bed}{The minimum time in bed in hours}

\item{col_names}{A list to override default column names. This function uses columns:
\itemize{
\item \code{time_in_bed}
}}
}
\value{
The sessions dataframe with only the sessions that meet the minimum time in bed requirement
}
\description{
Set minimum time in bed
}
\examples{
filtered_sessions <- set_min_time_in_bed(example_sessions, 2)
}
\seealso{
Other filtering: 
\code{\link{filter_by_night_range}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{remove_sessions_no_sleep}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
