% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_min_time_in_bed}
\alias{set_min_time_in_bed}
\title{Set minimum time in bed}
\usage{
set_min_time_in_bed(sessions, min_time_in_bed, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{min_time_in_bed}{The minimum time in bed in hours}

\item{return_mask}{If TRUE, return a logical vector indicating which sessions meet the minimum time in bed requirement}
}
\value{
The sessions dataframe with only the sessions that meet the minimum time in bed requirement, or a logical vector if \code{return_mask} is TRUE
}
\description{
Set minimum time in bed
}
\details{
This function uses columns:
\itemize{
\item \code{time_in_bed}
}
}
\examples{
filtered_sessions <- set_min_time_in_bed(example_sessions, 2)
}
\seealso{
Other filtering: 
\code{\link{filter_by_age_range}()},
\code{\link{filter_by_night_range}()},
\code{\link{filter_by_sex}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{remove_sessions_no_sleep}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_session_sleep_onset_range}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
