% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_session_sleep_onset_range}
\alias{set_session_sleep_onset_range}
\title{Set sleep onset time range}
\usage{
set_session_sleep_onset_range(sessions, from_time, to_time)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{from_time}{Include sessions where sleep started after this time (in format HH:MM)}

\item{to_time}{Include sessions where sleep started before this time (in format HH:MM)}
}
\value{
The sessions dataframe with only the sessions where sleep started within the specified time range
}
\description{
Set sleep onset time range
}
\examples{
filtered_sessions <- set_session_sleep_onset_range(example_sessions, "22:00", "06:00")
}
\seealso{
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}} to filter sessions based on start time.

Other filtering: 
\code{\link{filter_by_night_range}()},
\code{\link{filter_epochs_from_sessions}()},
\code{\link{remove_sessions_no_sleep}()},
\code{\link{select_devices}()},
\code{\link{select_subjects}()},
\code{\link{set_min_time_in_bed}()},
\code{\link{set_session_start_time_range}()}
}
\concept{filtering}
