% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_session_start_time_range}
\alias{set_session_start_time_range}
\title{Set session start time range}
\usage{
set_session_start_time_range(sessions, from_time, to_time)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{from_time}{Include sessions that started after this time (in format HH:MM)}

\item{to_time}{Include sessions that started before this time (in format HH:MM)}
}
\value{
The sessions dataframe with only the sessions that started within the specified time range
}
\description{
Set session start time range
}
