% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{shift_times_by_12h}
\alias{shift_times_by_12h}
\title{Shift times to break at 12 pm}
\usage{
shift_times_by_12h(times)
}
\arguments{
\item{times}{A vector of times in POSIXct format (or character convertible to POSIXct).}
}
\value{
A numerical vector of times (in hours) shifted to start at 12 PM
}
\description{
This function shifts times so that the day starts at 12 PM.
This is useful for plotting night data
}
\examples{
# Shift sessions start times to start at 12 PM
shifted_times <- shift_times_by_12h(example_sessions$session_start)

# Use dplyr::mutate to dicrectly add the shifted times to a dataframe
epochs <- example_epochs |>
  dplyr::mutate(shifted_time = shift_times_by_12h(timestamp))
}
\seealso{
Other time processing: 
\code{\link{convert_angle_to_time}()},
\code{\link{convert_times_to_mean_angle}()},
\code{\link{group_epochs_by_night}()},
\code{\link{group_sessions_by_night}()},
\code{\link{mean_time}()}
}
\concept{time processing}
