% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{social_jet_lag}
\alias{social_jet_lag}
\title{Calculate Social Jet Lag}
\usage{
social_jet_lag(
  sessions,
  midsleep = "time_at_midsleep",
  is_workday = "is_workday"
)
}
\arguments{
\item{sessions}{The sessions data frame}

\item{midsleep}{The name of the column containing mid-sleep times (default is "time_at_midsleep")}

\item{is_workday}{The name of the column indicating which days are workdays (default is "is_workday")}
}
\value{
The Social Jet Lag (SJL) value in hours
}
\description{
This function calculates the Social Jet Lag (SJL) metric as the difference in mid-sleep times
between workdays and free days.
}
\examples{
social_jet_lag(example_sessions)
}
\seealso{
Other sleep metrics: 
\code{\link{interdaily_stability}()}
}
\concept{sleep metrics}
